polyhaven2gmod/requirements.txt
requests==2.28.2
Pillow==9.5.0